Refactoring ot allow for T5 coding
