+++ src/main/java/com/ffe/traveller/classic/decoder/Planet.java
@@ -1,4 +1,4 @@
-package decoder;
+package com.ffe.traveller.classic.decoder;
 
 import java.util.Random;
