+++ src/main/java/com/ffe/traveller/classic/decoder/Starport.java
@@ -1,4 +1,4 @@
-package decoder;
+package com.ffe.traveller.classic.decoder;
 
 /**
  * @author markknights
