+++ src/main/java/com/ffe/traveller/classic/decoder/TradeClassifications.java
@@ -1,4 +1,4 @@
-package decoder;
+package com.ffe.traveller.classic.decoder;
 
 public enum TradeClassifications {
 	Agricultural , NonAgricultural , Industrial , NonIndustrial, Rich, Poor ;
