+++ src/main/resources/properties/classictraveller/speculative_trade.yml
