+++ src/main/resources/properties/classictraveller/universal_planetary_profile.yml
